guarantee_vehicle/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "guarantee-vehicle"
version = "0.1.0"
description = "Guarantee vehicle pricing and risk model"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "pyyaml>=6.0",
  "pydantic>=2.6",
  "matplotlib>=3.7",
  "openpyxl>=3.1",
  "streamlit>=1.36"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
